requirements/dev.txt
-r common.txt
django-debug-toolbar
django-extensions
django-template-repl
Werkzeug